pyproject.toml
[project]
name = "crypticreasoner-setter"
version = "0.1.0"
description = "An A2A agent that marks answers to Cryptic Crossword questions, and allows dictionary tool use"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "pydantic>=2.12.5",
    "uvicorn>=0.38.0",
    # This does not need dotenv / genai access
    # OTOH, the dictionary tool requires vector store...
    "fasttext==0.9.3",
    "huggingface_hub==1.3.1",
    "levenshtein==0.27.3"
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
